% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palpal.R
\docType{data}
\name{palpal_palettes}
\alias{palpal_palettes}
\title{List of available color-blind friendly palettes
export}
\format{An object of class \code{list} of length 2.}
\usage{
palpal_palettes
}
\description{
List of available color-blind friendly palettes
export
}
\keyword{datasets}
