% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palpal.R
\name{palpal}
\alias{palpal}
\title{Palette generator
Discrete only}
\usage{
palpal(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of palette: \code{wong} or \code{mkweb}}

\item{n}{Number of colors, default all}

\item{type}{Only "discrete" is enabled for the now}
}
\value{
vector of colors
}
\description{
Palette generator
Discrete only
}
\examples{
palpal("wong")
palpal("mkweb")
\dontrun{ggplot(data=mtcars, aes(x=mpg, y=disp, color=factor(cyl))) +
geom_point() + scale_color_manual(values=palpal("mkweb"))}
}
